Move utils.h.
