+++ cpp/common/utils.h
