+++ cpp/serving/sparse_lookup_model.cpp
@@ -19,7 +19,7 @@
 #include <serving/inmem_sparse_lookup.h>
 #include <serving/sparse_lookup_model.h>
 #include <serving/threadpool.h>
-#include <serving/utils.h>
+#include <common/utils.h>
 
 #include <boost/core/demangle.hpp>
 #include <fmt/format.h>
