+++ cpp/tests/serving/feature_compute_funcs_test.cpp
@@ -16,7 +16,7 @@
 
 #include <common/hash_utils.h>
 #include <serving/test_utils.h>
-#include <serving/utils.h>
+#include <common/utils.h>
 
 #include <arrow/api.h>
 #include <arrow/compute/api.h>
